backend-relay/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.134.0",
    "nats-py>=2.9.0",
    "python-dotenv>=1.2.1",
    "uvicorn[standard]>=0.41.0",
    "websockets>=16.0",
]
